Import ancient version of MIC-1/MAC-1 virtual machine
I wrote this code years and years ago (between 2003 and 2005) for a
Computer Architecture programming project in highschool (at TJHSST).  It
is an implementation of the MIC-1 microarchitecture and MAC-1 virtual
machine as specified by the "Structured Computer Organization" book;
whatever version of it I happened to have at the time.

It's only about a thousand lines of perl, but it even manages to include
a fully-functional micro-assembler for the MIC-1.

Unfortunately the command-line syntax is crap and it's missing all kinds
of useful features like binary-output, etc, but it is capable of loading
a MAC-1 program and then stepping through each MIC-1 microarchitectural
operation.

Given that I was 16 or 17 when I wrote this, it is probably horribly
full of bugs, but at the same time I feel a bit proud of how clean and
well commented the code is for a microcode interpreter (especially given
that it is written in Perl!)

Signed-off-by: Kyle Moffett <kyle@moffetthome.net>